requirements.txt
gdeltdoc~=1.5.0
pandas~=2.2.2
newspaper3k
lxml[html_clean]
lxml_html_clean
nltk
PyYAML~=6.0.1
requests~=2.32.3
pymongo~=4.8.0
confluent-kafka